+++ gframe/single_duel.cpp
@@ -1565,7 +1565,7 @@ void SingleDuel::RefreshSingle(int player, int location, int sequence, int flag)
 			NetServer::ReSendToPlayer(*pit);
 	}
 }
-int SingleDuel::MessageHandler(long fduel, int type) {
+int SingleDuel::MessageHandler(intptr_t fduel, int type) {
 	if(!enable_log)
 		return 0;
 	char msgbuf[1024];
